+++ src/network/networkMonitor.js
@@ -171,7 +171,7 @@ export class NetworkMonitor extends Monitor {
                 }.bind(this))
                 .catch(e => {
                     if(e.isCancelled) {
-                        Utils.log('Update canceled: ' + key);
+                        Utils.log('Network Monitor update canceled: ' + key);
                     }
                     else {
                         Utils.error(e.message);